category, search and delete

+++ modules/control.js
@@ -26,14 +26,21 @@ const modalClose = (overlay) => {
   modalForm.reset();
 };
 
-export const addProduct = (overlay, modalForm, table, total, addGoods) => {
+export const addProduct = (overlay, modalForm, total, addGoods) => {
   addGoods.addEventListener("click", () => {
     modalOpen(overlay, modalForm);
     controlCheckbox(modalCheckbox, modalInputDiscount);
+    total.textContent = "";
+    console.log(total);
   });
+  getCategory();
   uploadFile();
   // Итоговая стоимость в модальном окне должна правильно высчитываться при смене фокуса
   modalForm.price.addEventListener("change", (e) => {
+    const total = modalForm.price.value * modalForm.count.value;
+    modalForm.total.textContent = total;
+  });
+  modalForm.count.addEventListener("change", (e) => {
     const total = modalForm.price.value * modalForm.count.value;
     modalForm.total.textContent = total;
   });
@@ -50,23 +57,30 @@ export const deleteRow = (list) => {
   list.addEventListener("click", (e) => {
     const target = e.target;
     if (target.closest(".table__btn_del")) {
-      const tr = target.closest("tr");
-      const id = tr.querySelector(".table__cell-id").textContent.slice(4);
-      console.log(id);
-      fetchRequest(`goods/${id}`, {
-        method: "DELETE",
+      const confirm = document.querySelector(".confirm__overlay");
+      const agreeBtn = confirm.querySelector(".modal__submit");
+      const closeBtn = confirm.querySelector(".modal__cancel");
+      confirm.classList.add("active");
+      confirm.addEventListener("click", (e) => {
+        e.preventDefault();
+        if (e.target === agreeBtn) {
+          const tr = target.closest("tr");
+          const id = tr.querySelector(".table__cell-id").textContent.slice(4);
+          console.log(id);
+          tr.remove();
+          fetchRequest(`goods/${id}`, {
+            method: "DELETE",
+          });
+          modalClose(confirm);
+        } else if (
+          e.target.closest(".confirm__overlay") ||
+          e.target.closest(".modal__close") ||
+          e.target === closeBtn
+        ) {
+          modalClose(confirm);
+        }
       });
-
-      // arr.splice(id, 1);
-      // tr.remove();
-      // renderGoods(arr, table);
-      // const tableSum = getTableSum();
-      // const totalprice = totalUpdate(total, tableSum);
-      // total.textContent = ` $ ${totalprice}`;
     }
-    // fetchRequest("goods", {
-    //   callback: renderGoods,
-    // });
   });
 };
 export const totalUpdate = (total, arr) => {
@@ -93,26 +107,43 @@ const controlCheckbox = (modalCheckbox, modalInputDiscount) => {
   });
 };
 
-const submitProduct = (modalForm, overlay) => {
+export const submitProduct = (modalForm, overlay) => {
   modalForm.addEventListener("submit", (e) => {
     e.preventDefault();
-    // const id = getVendorCode();
-    // console.log(id);
-    const formData = new FormData(e.target);
-    // formData.append("id", id.textContent);
-    const newProduct = Object.fromEntries(formData);
-    console.log(newProduct);
-    const resp = fetchRequest("goods", {
-      method: "POST",
-      body: newProduct,
-      callback: (err, data) => {
-        if (err) {
-          console.warn(err, data);
-          return;
-        }
-      },
-    });
-    console.log(resp);
+
+    const btn = document.querySelector(".modal__submit").textContent;
+    console.log(btn);
+    if (btn === "Добавить товар") {
+      const formData = new FormData(e.target);
+      const newProduct = Object.fromEntries(formData);
+      const resp = fetchRequest("goods", {
+        method: "POST",
+        body: newProduct,
+        callback: (err, data) => {
+          if (err) {
+            console.warn(err, data);
+            return;
+          }
+        },
+      });
+      console.log(resp);
+    } else {
+      const formData = new FormData(e.target);
+      const newProduct = Object.fromEntries(formData);
+      console.log(newProduct);
+      const resp = fetchRequest("goods/", {
+        method: "PATCH",
+        body: newProduct,
+        callback: (err, data) => {
+          if (err) {
+            console.warn(err, data);
+            return;
+          }
+        },
+      });
+      console.log(resp);
+    }
+
     modalForm.reset();
     modalClose(overlay);
     fetchRequest("goods", {
@@ -144,4 +175,25 @@ const uploadFile = () => {
       }
     }
   });
+};
+const datalist = document.querySelector("#category-list");
+const getCategory = () => {
+  const data = fetchRequest("category", {
+    method: "GET",
+    callback: createCategory,
+  });
+  console.log(data);
+};
+const createCategory = (data, err) => {
+  if (err) {
+    console.warn(err, data);
+    return;
+  }
+  const categories = data.map((elem) => {
+    const option = document.createElement("option");
+    console.log("elem", elem);
+    option.value = elem;
+    return option;
+  });
+  datalist.append(...categories);
 };
